% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlnt-package.R
\docType{package}
\name{cvlnt-package}
\alias{cvlnt}
\alias{cvlnt-package}
\title{cvlnt: R Interface To Covalent HQ API}
\description{
cvlnt is a third-party interface for the Covalent HQ API. The
    Covalent HQ API lets you retrieve data from the Ethereum blockchain.
}
\author{
\strong{Maintainer}: Santiago Gualchi \email{santiagogualchi@gmail.com}

}
\keyword{internal}
