% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_changes_in_holders.R
\name{get_changes_in_holders}
\alias{get_changes_in_holders}
\title{Get Changes In Token Holders Between Two Block Heights}
\usage{
get_changes_in_holders(
  token_contract,
  api_key,
  starting_block,
  ending_block = "latest",
  sleep = 1/10,
  ...
)
}
\arguments{
\item{token_contract}{The token contract address.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{starting_block}{Starting block to define a block range.}

\item{ending_block}{Ending block to define a block range.}

\item{sleep}{Interval in seconds between requests.}

\item{...}{Additional arguments to pass to the request. For more information
on available parameters see
\href{https://www.covalenthq.com/docs/api/}{Covalent HQ API}.}
}
\value{
Returns a data.frame with the current/historical balances for the specified
token contract of all holders.
}
\description{
Get Changes In Token Holders Between Two Block Heights
}
\examples{
\dontrun{
contract <- "0x0954906da0Bf32d5479e25f46056d22f08464cab"

get_changes_in_holders(contract, MY_API_KEY, starting_block = "11500000")

get_changes_in_holders(contract, MY_API_KEY, starting_block = "11500000",
                       ending_block = "11550000")
}
}
