% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Performs A Request To A Covalent HQ Endpoint}
\usage{
get_endpoint(endpoint, api_key, ...)
}
\arguments{
\item{endpoint}{The endpoint's URI.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{...}{Additional arguments to pass to the request. For more information
on available parameters see
\href{https://www.covalenthq.com/docs/api/}{Covalent HQ API}.}
}
\value{
Returns the endpoint data for the requested page.
}
\description{
Performs A Request To A Covalent HQ Endpoint
}
\examples{
\dontrun{
endpoint <- paste0("https://api.covalenthq.com/v1/1/address/",
                   "0xe6a7a1d47ff21b6321162aea7c6cb457d5476bca",
                   "/balances_v2/")

get_endpoint(endpoint, MY_API_KEY)
get_endpoint(endpoint, MY_API_KEY, "block-height" = "11601973")
}
}
