% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_erc20_transfers.R
\name{get_erc20_transfers}
\alias{get_erc20_transfers}
\title{Get ERC20 Token Transfers}
\usage{
get_erc20_transfers(address, token_contract, api_key, sleep = 1/10, ...)
}
\arguments{
\item{address}{The wallet address.}

\item{token_contract}{The ERC20 token contract address to retrieve transfers
of.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{sleep}{Interval in seconds between requests.}

\item{...}{Additional arguments to pass to the request. For more information
on available parameters see
\href{https://www.covalenthq.com/docs/api/}{Covalent HQ API}.}
}
\value{
Returns a data.frame with the transfers of the specified ERC20 token for the
specified address.
}
\description{
Get ERC20 Token Transfers
}
\examples{
\dontrun{
address <- "0x669184d1fe949249857439232ee67d7adb25d0ae"
token_contract <- "0x0954906da0Bf32d5479e25f46056d22f08464cab"

get_erc20_transfers(address, token_contract, MY_API_KEY)
get_erc20_transfers(address, token_contract, MY_API_KEY,
                    "quote-currency" = "eth")
}
}
