% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_block.R
\name{get_latest_block}
\alias{get_latest_block}
\title{Get Latest Block}
\usage{
get_latest_block(api_key)
}
\arguments{
\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}
}
\value{
A character with the latest block height.
}
\description{
Get Latest Block
}
\examples{
\dontrun{
get_latest_block(MY_API_KEY)
}
}
