% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page.R
\name{get_page}
\alias{get_page}
\title{Performs a Request to Get a Specific Page of an Endpoint}
\usage{
get_page(endpoint, api_key, page_number, ...)
}
\arguments{
\item{endpoint}{The endpoint's URI.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{page_number}{The paginator's page number parameter.}

\item{...}{Additional arguments to pass to the request.}
}
\value{
Returns the endpoint data for the requested page.
}
\description{
Performs a Request to Get a Specific Page of an Endpoint
}
\examples{
\dontrun{
endpoint <- paste0("https://api.covalenthq.com/v1/1/tokens/",
                   "0x0954906da0Bf32d5479e25f46056d22f08464cab",
                   "/token_holders/")

get_page(endpoint, MY_API_KEY, 12L)
get_page(endpoint, MY_API_KEY, 12L, "block-height" = "11601973")
}
}
