% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paginated_endpoint.R
\name{get_paginated_endpoint}
\alias{get_paginated_endpoint}
\title{Get Paginated Data From An Endpoint}
\usage{
get_paginated_endpoint(endpoint, api_key, sleep = 1/10, ...)
}
\arguments{
\item{endpoint}{The endpoint's URI.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{sleep}{Interval in seconds between requests.}

\item{...}{Additional arguments to pass to the request. For more information
on available parameters see
\href{https://www.covalenthq.com/docs/api/}{Covalent HQ API}.}
}
\value{
Returns the data for the requested endpoint as a data.frame.
}
\description{
Get Paginated Data From An Endpoint
}
\examples{
\dontrun{
endpoint <- paste0("https://api.covalenthq.com/v1/1/tokens/",
                   "0x0954906da0Bf32d5479e25f46056d22f08464cab",
                   "/token_holders/")

get_paginated_endpoint(endpoint, MY_API_KEY)
get_paginated_endpoint(endpoint, MY_API_KEY, "block-height" = "11601973")
}
}
