% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simple_endpoint.R
\name{get_simple_endpoint}
\alias{get_simple_endpoint}
\title{Get Unpaginated Data from an Endpoint}
\usage{
get_simple_endpoint(endpoint, api_key, ...)
}
\arguments{
\item{endpoint}{The endpoint's URI.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{...}{Additional arguments to pass to the request.}
}
\value{
Returns the data for the requested endpoint as a data.frame.
}
\description{
Get Unpaginated Data from an Endpoint
}
\examples{
\dontrun{
endpoint <- paste0("https://api.covalenthq.com/v1/1/address/",
                   "0xe6a7a1d47ff21b6321162aea7c6cb457d5476bca",
                   "/balances_v2/")

get_simple_endpoint(endpoint, MY_API_KEY)
get_simple_endpoint(endpoint, MY_API_KEY, "quote-currency" = "eth")
}
}
