% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transaction.R
\name{get_transaction}
\alias{get_transaction}
\title{Get A Transaction}
\usage{
get_transaction(tx_hash, api_key, ...)
}
\arguments{
\item{tx_hash}{The transaction hash.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{...}{Additional arguments to pass to the request. For more information
on available parameters see
\href{https://www.covalenthq.com/docs/api/}{Covalent HQ API}.}
}
\value{
Returns a data.frame with a transaction's data.
}
\description{
Get A Transaction
}
\examples{
\dontrun{
tx_hash <- "0xbda92389200cadac424d64202caeab70cd5e93756fe34c08578adeb310bba254"

get_transaction(tx_hash, MY_API_KEY)
}
}
