% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transactions.R
\name{get_transactions}
\alias{get_transactions}
\title{Get Transactions}
\usage{
get_transactions(address, api_key, chain_id = "1", sleep = 1/10, ...)
}
\arguments{
\item{address}{The wallet address.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{chain_id}{Chain ID of the Blockchain being queried. Supports "137" for
Matic, defaults to "1" for Ethereum Mainnet.}

\item{sleep}{Interval in seconds between requests.}

\item{...}{Additional arguments to pass to the request. For more information
on available parameters see
\href{https://www.covalenthq.com/docs/api/}{Covalent HQ API}.}
}
\value{
Returns all transactions for \code{address} including their decoded log events as
a list on the \code{log_events} column.
}
\description{
Get Transactions
}
\examples{
\dontrun{
address <- "0x5a6d3b6bf795a3160dc7c139dee9f60ce0f00cae"

get_transactions(address, MY_API_KEY)
get_transactions(address, MY_API_KEY, "quote-currency" = "eth")
}
}
