% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_params.R
\name{parse_params}
\alias{parse_params}
\title{Parses Named Parameters As A Query String}
\usage{
parse_params(params)
}
\arguments{
\item{params}{A named list of arguments.}
}
\value{
Returns an atomic character with the parameters as a query string.
}
\description{
Parses Named Parameters As A Query String
}
\examples{
\dontrun{
x <- parse_params(list("page-number" = "1", "block-height" = "10000"))
x == "page-number=1&block-height=10000" # TRUE.
}
}
