% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlnt-package.R
\docType{package}
\name{cvlnt-package}
\alias{cvlnt}
\alias{cvlnt-package}
\title{cvlnt: R Interface To Covalent HQ API}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Santiago Gualchi \email{santiagogualchi@gmail.com}

}
\keyword{internal}
