% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nft_metadata.R
\name{get_nft_metadata}
\alias{get_nft_metadata}
\title{Get External NFT Metadata}
\usage{
get_nft_metadata(address, token_id, api_key, ...)
}
\arguments{
\item{address}{An ERC751 or ERC1155 NFT contract address.}

\item{token_id}{The NFT token ID.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{...}{Additional arguments to pass to the request.}
}
\value{
Returns a data.frame with an NFT external metadata.
}
\description{
Get External NFT Metadata
}
\examples{
\dontrun{
address <- "0xe4605d46fd0b3f8329d936a8b258d69276cba264"
token_id <- "123"

get_nft_metadata(address, token_id, MY_API_KEY)
}
}
