% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token_holders.R
\name{get_token_holders}
\alias{get_token_holders}
\title{Get Data of a Token's Holders}
\usage{
get_token_holders(token_contract, api_key, ..., sleep = 1/10)
}
\arguments{
\item{token_contract}{The token contract address.}

\item{api_key}{The user's public API key. To generate it, you need an account
at \href{https://www.covalenthq.com/}{Covalent HQ}.}

\item{...}{Additional arguments to pass to the request.}

\item{sleep}{Interval in seconds between requests.}
}
\value{
Returns a data.frame with the balance for the specified token contract of all
holders.
}
\description{
Get Data of a Token's Holders
}
\examples{
\dontrun{
tc <- "0x0954906da0Bf32d5479e25f46056d22f08464cab"

get_token_holders(tc, MY_API_KEY)
get_token_holders(tc, MY_API_KEY, "block-height" = "11601973")
}
}
