% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_params.R
\name{parse_params}
\alias{parse_params}
\title{Parses Named Parameters as a Query String}
\usage{
parse_params(...)
}
\arguments{
\item{...}{Named arguments}
}
\value{
Returns an atomic character with the parameters as a query string.
}
\description{
Parses Named Parameters as a Query String
}
\examples{
\dontrun{
x <- parse_params("page-number" = "1", "block-height" = "10000")
x == "page-number=1&block-height=10000" # TRUE.
}
}
